+++ weblate/machinery/base.py
@@ -187,10 +187,9 @@ class MachineTranslation(object):
 
     def report_error(self, exc, message):
         """Wrapper for handling error situations"""
-        extra = {'mt_url': self.request_url, 'mt_params': self.request_params}
-        report_error(exc, extra_data=extra, prefix='Machinery error')
+        report_error(exc, prefix='Machinery error')
         LOGGER.error(message, self.name)
-        LOGGER.info('Last URL: %s, params: %s', extra['mt_url'], extra['mt_params'])
+        LOGGER.info('Last URL: %s, params: %s', self.request_url, self.request_params)
 
     def get_supported_languages(self):
         """Return list of supported languages."""